core refacto and test for cner model
